+++ HISTORY.rst
@@ -24,6 +24,7 @@ Unreleased
 * ALCN-161: Modify the procurement wizard to allow the user to select a specific day or a supplier
 * ALCN-161: Add module stock_orderpoint_product
 * ALCN-1069: Add import min/max
+* ALCN-1204: Add Cp2z02 (fardelage) in product xml (wso2)
 
 **Bugfixes**